Feat/company valuation: Added the Employees and contracts router, and the Shares Calculation file.

+++ migrations/20220702161910-create-company.js
@@ -14,7 +14,7 @@ module.exports = {
         defaultValue: 'myCompany BV'
       },
       currentValuation: {
-        type: Sequelize.NUMERIC(12, 2),
+        type: Sequelize.FLOAT,
         allowNull: false
       },
       totalCompanyShares: {
